// src/shared/model/types/invoiceStructure/results/medicalCases/GetCompletedCaseListItemsResult.ts
export interface GetCompletedCaseListItemsResult {
  completedCases: CompletedCaseListItemDto[];
  totalCount: number;
}

export interface CompletedCaseListItemDto {
  completedCaseUid: number;
  entryNumber: number;
  amountBilled: number;
  approvedAmount: number | null;
  insuranceCompanyApprovedAmount: number | null;
  medicalCareConditions: number;
  patientLastName: string;
  patientFirstName: string;
  patientMiddleName: string;
  insurancePolicySeries: string | null;
  insurancePolicyNumber: string;
  entryPositionNumber: number;
}